filesystem: added hash_filesystem function
The hash_filesystem function returns a dictionary containing the path
and hash of each file contained within the File System.

It tries to hash the entire disk at first, in case of errors it
iterates over all the nodes logging problematic ones.

This allows to retrieve the content of filesystems in inconsistent
state due to errors or live snapshotting.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>